Oh boy, lots of stuff. Sourced and migrated to another game fork.
Had another fork of this project laying in an un-staged folder for some
testing. Only included some fixes to the game window for better mobile
experience. It did also bring with it some bugs and bad refactors. Give
some take some.

Fixed a mousepointer bug that I think was present in the original as
well.